// chapter-08/src/types/generated/types.gen.ts
export type UserSummary = {
  id: string;
  email: string;
  username: string;
};

export type Idea = {
  id: string;
  title: string;
  shortDescription: string;
  description: string;
  tags: Array<string>;
  authorId: string;
  author: UserSummary;
  reviewsCount: number;
  avgRating: number | null;
  createdAt: string;
  updatedAt: string;
};

export type Pagination = {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
  prevPage: number | null;
  nextPage: number | null;
};

export type IdeaListWithPaginationResponse = {
  data: Array<Idea>;
  pagination: Pagination;
};

export type IdeaListResponse = {
  data: Array<Idea>;
};

export type IdeaSummary = {
  id: string;
  title: string;
};

export type Review = {
  id: string;
  content: string;
  rating: number;
  authorId: string;
  author: UserSummary;
  ideaId: string;
  idea?: IdeaSummary;
  createdAt: string;
  updatedAt: string;
};

export type ReviewListResponse = {
  data: Array<Review>;
};